+++ src/app/dashboard/companies/page.tsx
